requirements.txt
rpaframework
pytesseract
pillow